Made new teams file to display teams and got rid of favoriteTeams
