+++ app/(tabs)/login.tsx
@@ -30,7 +30,8 @@ export default function LoginScreen() {
     setLoading(true);
   
     try {
-      const response = await fetch('http://192.168.1.215:8080/api/login', { // points to back end so back end has to fetch from database ip might be hardcoded?
+      const response = await fetch('http://10.11.140.56:8080/users/login', {
+      // const response = await fetch('http://192.168.1.215:8080/api/login', { // points to back end so back end has to fetch from database ip might be hardcoded?
         method: "POST", // SENDING TO BACKEND / POST MAPPING
         headers: {
           "Content-Type": "application/json",
@@ -49,7 +50,7 @@ export default function LoginScreen() {
         await AsyncStorage.setItem("username", data.userName);
 
         Alert.alert("Welcome", `Hello, ${data.userName}!`);
-        navigation.navigate("FavoriteTeams", { username: data.userName });
+        navigation.navigate("teams", { username: data.userName });
       } else {
         Alert.alert("Error", data.message || "Login failed.");
       }
